refactor: reorganize team wallet feature layering
