+++ src/app/[locale]/(back-office)/team/[teamId]/team-wallet/_components/team-wallet/charge-cards-tab.tsx
@@ -3,7 +3,8 @@
 import { Button } from '@/components/ui/button'
 import { Input } from '@/components/ui/input'
 import { Plus } from 'lucide-react'
-import { ChargeCard } from '../../_schemas/team-wallet-types'
+
+import { ChargeCard } from '../../_schemas/team-wallet.schema'
 import { ChargeCardsTable } from './charge-cards-table'
 
 interface ChargeCardsTabProps {
