+++ src/app/[locale]/(back-office)/team/[teamId]/team-wallet/_components/team-wallet/charge-sessions-table.tsx
@@ -1,7 +1,8 @@
 'use client'
 
 import { Button } from '@/components/ui/button'
-import { ChargeSession } from '../../_schemas/team-wallet-types'
+
+import { ChargeSession } from '../../_schemas/team-wallet.schema'
 
 interface ChargeSessionsTableProps {
   sessions: ChargeSession[]
