+++ src/app/[locale]/(back-office)/team/[teamId]/team-wallet/_layers/presentation/index.ts
@@ -1,0 +1,1 @@
+export { TeamWalletPage } from './team-wallet-page'
